Add pickblock for structure preview
Add the ability to pickblock block that are in a structure preview

Note: the pickblock will *only* work on the currently previewed block, and not any compatible block.
eg. if in survival and the position you pickblock allows for either stone or dirt
if the preview is showing stone, and you have dirt in your inventory and pickblock it, then it will not select the dirt.

Signed-off-by: solonovamax <solonovamax@12oclockpoint.com>